requirements.txt
flask
spacy
requests
sentence-transformers
faiss-cpu
transformers
torch
numpy
